chore(refactoring): created a separate function for each controller
also adjusted folder structure and splitted tests as well

closes #13
