+++ src/_shared/types/contoller.ts
@@ -1,0 +1,3 @@
+import { PluginMessage } from './actions';
+
+export type Controller = (msg: PluginMessage) => void;
