--- src/_shared/types/resolver.ts
@@ -1,3 +1,0 @@
-import { PluginMessage } from './actions';
-
-export type Resolver = (msg: PluginMessage) => void;
