+++ src/plugin/closePlugin/closePlugin.ts
@@ -1,0 +1,5 @@
+import { Controller } from '../../_shared/types/contoller';
+
+export const closePlugin: Controller = () => {
+  figma.closePlugin();
+};
