+++ src/plugin/controller.ts
@@ -1,4 +1,16 @@
-import { controllers } from './pluginControllers';
+import { PluginActionTypes, ActionTypes } from '../_shared/types/actions';
+import { Controller } from '../_shared/types/contoller';
+import { generateTheme } from './generateTheme/generateTheme';
+import { syncTheme } from './syncTheme/syncTheme';
+import { createMorfeoTheme } from './createMorfeoTheme/createMorfeoTheme';
+import { closePlugin } from './closePlugin/closePlugin';
+
+export const controllers: Record<PluginActionTypes, Controller> = {
+  [ActionTypes.generateTheme]: generateTheme,
+  [ActionTypes.syncTheme]: syncTheme,
+  [ActionTypes.createMorfeoTheme]: createMorfeoTheme,
+  [ActionTypes.closePlugin]: closePlugin,
+};
 
 figma.showUI(__html__);
